+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Can now actually delete todos.
+
 ## [1.4.2] - 2019-12-15
 ### Added
 - `Macroy.delete_todo/1`.